Update toolkit.py
- added a few languages of month names
- corrected (?) the field sequence in date formats
- set some date formats to have either 2- or 4-digit years
- insert a cleanup line for separators enclosed between spaces (' : ')
note, the 2-digit years may cause malidentification if the time string is not reliably separated. hence the cleanup. still, two-digit years may require 19- or 20- usually; i suggest adding an option to set the century: if no century is set, DATEFORMATS=DATEFORMATS.replace('{2,4}','{4}'), otherwise, if a 2-digit year is found, the set century is used to prefix it. but this option goes beyond the present file so i did not implement it.