Refactor Project-Articleset relation
Adds the model ProjectArticleset as 'through' model for the
Project.articlesets field.

Removes the Project.favourite_articlesets many to many field and replaces
it with the ProjectArticleset.is_favourite boolean field.

Articlesets must be included in the project's `articlesets` field in order for
them to be considered part of the project. Ownership through the
`ArticleSet.project` field now only has influence on the administration
permissions.